edit mutations and styling updates
