+++ src/graphql/mutations/ambulances.ts
@@ -1,0 +1,18 @@
+import gql from 'graphql-tag';
+
+export const ADD_AMBULANCE = gql`
+  mutation addAmbulance($vehicleNumber: Int!) {
+    addAmbulance(vehicleNumber: $vehicleNumber) {
+      vehicleNumber
+    }
+  }
+`;
+
+export const EDIT_AMBULANCE = gql`
+  mutation updateAmbulance($id: ID!, $vehicleNumber: Int!) {
+    updateAmbulance(id: $id, vehicleNumber: $vehicleNumber) {
+      id
+      vehicleNumber
+    }
+  }
+`;
