--- src/graphql/mutations/templates/ambulances.ts
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag';
-
-export const ADD_AMBULANCE = gql`
-  mutation addAmbulance($vehicleNumber: Int!) {
-    addAmbulance(vehicleNumber: $vehicleNumber) {
-        vehicleNumber
-    }
-  }
-`;
