--- src/graphql/mutations/templates/hospitals.ts
@@ -1,9 +1,0 @@
-import gql from 'graphql-tag';
-
-export const ADD_HOSPITAL = gql`
-  mutation addHospital($name: String!) {
-    addHospital(name: $name) {
-      name
-    }
-  }
-`;
